fix(prep): cut one-hot dummy memory 8× — int64 → uint8 (GMR-0022 OOM)
GMR-0022 prep crashed 2026-05-22 during cohort resolution at
src/features.py:488 with numpy OOM:

  numpy._core._exceptions._ArrayMemoryError: Unable to allocate
  58.9 GiB for an array with shape (825248, 9586) and data type int64

Root cause: `pd.get_dummies(exploded, dtype=int)` materializes a dense
int64 matrix sized (n_exploded_rows × n_unique_values). For
binary 0/1 data int64 wastes 7 of every 8 bytes. GMR-0022's broad
software-developer cohort (~64k workers exploded to 825k rows across
9.5k unique skill tags) pushed the allocation past the 14 GiB GHA
runner limit.

Switching to dtype="uint8" (1 byte/cell) cuts the same allocation to
~7.4 GiB — comfortably under runner RAM. Downstream code is unchanged
(groupby.max + reindex(fill_value=0) both work on uint8).

Knock-on effects of the GMR-0022 OOM (not fixed in this commit, noted
for follow-up):
  - Prep aborted before reaching the ICP enrichment block at
    main.py:2487, so cohort_icp was never persisted for GMR-0022.
    Console shows no ICP card.
  - Smart-ramp poller cached GMR-0022's signature in
    processed_ramps.json after the (failed-but-returning) prep run.
    Every subsequent poll sees the unchanged signature and skips
    GMR-0022. Operator must force re-process via workflow_dispatch
    with ramp_id=GMR-0022 to refill cohort_icp + briefs.

Tested: 170 of 171 tests pass; the 1 failure (test_meta_targeting
flexible_spec assertion) is pre-existing — SPECIAL_AD_CATEGORY=EMPLOYMENT
strips flexible_spec in MetaInterestResolver per
feedback_outlier_meta_tiktok_gmr0021.

+++ src/features.py
@@ -484,8 +484,13 @@ def binary_features(df: pd.DataFrame, freq_maps: dict[str, Counter]) -> pd.DataF
         if exploded.empty:
             continue
 
-        # get_dummies on the exploded series → binary matrix with original index
-        dummies = pd.get_dummies(exploded, dtype=int)
+        # get_dummies on the exploded series → binary matrix with original index.
+        # dtype="uint8" (1 byte/cell) instead of int=int64 (8 bytes/cell). For
+        # binary one-hot data the wider dtype just wastes memory. Surfaced
+        # 2026-05-22 when GMR-0022 prep OOM'd with int64: 825248 × 9586 cells
+        # = 58.9 GiB requested allocation, refused by numpy on a 14 GiB runner.
+        # uint8 cuts that to ~7.4 GiB, fits comfortably in standard GHA runners.
+        dummies = pd.get_dummies(exploded, dtype="uint8")
 
         # aggregate back to original row index (any occurrence = 1)
         dummies = dummies.groupby(level=0).max()
